Update tech_assignment.md
I removed a few sentences that reiterated previously stated thoughts as these sentences didn't improve the flow of information nor did they provide clarity. Additionally, I wonder if it would be better to reference git merge, as "git merge" and not mention "origin/main"? I bring this up because the topic of git merge is beyond the scope of this article and adding this additional information adds another layer of complexity that could be avoided if we remove it.  -M